Hackathon-2-spec-Driven-Development/Phase-1-Todo-console-App/pyproject.toml
[project]
name = "todo-app"
version = "0.2.0"
description = "Phase I Todo Console Application with Advanced Features"
requires-python = ">=3.12"
dependencies = [
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
